// src/types.ts
export type Meta = {
  title: string;
  description: string;
};

export type NavItem = {
  label: string;
  href: string;
};

export type IconName =
  | 'building'
  | 'calendar'
  | 'card'
  | 'certificate'
  | 'clock'
  | 'document'
  | 'download'
  | 'education'
  | 'finance'
  | 'globe'
  | 'grid'
  | 'mail'
  | 'map'
  | 'person'
  | 'phone'
  | 'search'
  | 'shield'
  | 'structure'
  | 'support'
  | 'table'
  | 'target'
  | 'unlock';